# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Web API для управления шахматными партиями и сущностями (игроки, фигуры, ходы).",
        "title": "Chess Game API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/figure": {
            "post": {
                "description": "Создает новую фигуру",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "figures"
                ],
                "summary": "Создать фигуру",
                "parameters": [
                    {
                        "description": "Данные фигуры",
                        "name": "figure",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.figureDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
        "/figure/{id}": {
            "get": {
                "description": "Возвращает фигуру по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "figures"
                ],
                "summary": "Получить фигуру",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фигуры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example_com_go-basic_packages_internal_models_chess.Figure"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет фигуру по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "figures"
                ],
                "summary": "Обновить фигуру",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фигуры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "figure",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.figureDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет фигуру по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "figures"
                ],
                "summary": "Удалить фигуру",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фигуры",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
...
        "/move": {
            "post": {
                "description": "Создает новый ход в хранилище",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Создать ход",
                "parameters": [
                    {
                        "description": "Данные хода",
                        "name": "move",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.moveDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
        "/move/{id}": {
            "get": {
                "description": "Возвращает ход по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Получить ход",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID хода",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example_com_go-basic_packages_internal_models_player.Move"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет ход по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Обновить ход",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID хода",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные хода",
                        "name": "move",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.moveDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет ход по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Удалить ход",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID хода",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
...
        "/player": {
            "post": {
                "description": "Создает нового игрока в хранилище",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Создать игрока",
                "parameters": [
                    {
                        "description": "Данные игрока",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.playerDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
        "/player/{id}": {
            "get": {
                "description": "Возвращает игрока по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Получить игрока",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игрока",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example_com_go-basic_packages_internal_models_player.Player"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные игрока по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Обновить игрока",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игрока",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.playerDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет игрока по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Удалить игрока",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игрока",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/packages_internal_handlers.Err"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "example_com_go-basic_packages_internal_models_chess.Figure": {
            "type": "object",
            "properties": {
                "game_color": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "piece_color": {
                    "type": "string"
                },
                "symbol": {
                    "type": "integer"
                }
            }
        },
        "example_com_go-basic_packages_internal_models_player.Move": {
            "type": "object",
            "properties": {
                "figure": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "position_from": {
                    "type": "string"
                },
                "position_to": {
                    "type": "string"
                },
                "time_took": {
                    "$ref": "#/definitions/time.Duration"
                }
            }
        },
...
        "packages_internal_handlers.Err": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                }
            }
        },
...
        "packages_internal_handlers.figureDTO": {
            "type": "object",
            "required": [
                "game_color",
                "piece_color",
                "symbol"
            ],
            "properties": {
                "game_color": {
                    "type": "string"
                },
                "piece_color": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
        "packages_internal_handlers.moveDTO": {
            "type": "object",
            "required": [
                "figure",
                "position_from",
                "position_to"
            ],
            "properties": {
                "figure": {
                    "type": "string"
                },
                "position_from": {
                    "type": "string"
                },
                "position_to": {
                    "type": "string"
                },
                "time_took": {
                    "type": "integer"
                }
            }
        },
...
    }
}